requirements.txt
Flask
google-generativeai
langchain
langchain-google-genai
Pillow
python-dotenv